+++ commons/src/androidMain/kotlin/test/TestBalloonCommons.android.kt
@@ -1,4 +1,3 @@
 package at.asitplus.testballoon
 
-actual val String.escaped: String get() = this //= replace('/', '⧸')
-actual  val DEFAULT_TEST_NAME_MAX_LEN : Int = 10
+actual  val defaultMaxLen : Int = 10