+++ commons/src/nonAndroidMain/kotlin/at/asitplus/test/TestBalloonCommons.nonAndroid.kt
@@ -1,4 +1,3 @@
 package at.asitplus.testballoon
 
-actual val String.escaped: String get() = this
-actual val DEFAULT_TEST_NAME_MAX_LEN : Int = 64
+actual val defaultMaxLen : Int = 64